device-server: use FileHandler instead of RotatingFileHandler for logs.
Automatically rotating the logs, with RotatingFileHandler could be
useful.  However, we are not actually rotating anything because we are
not setting its maxBytes or backupCount values and by default it never
rotates.  We might as well use the standard FileHandler and come back
to RotatingFileHandler later when we actually configure it.

(cherry picked from commit 2c4f10c0df65ec971665f0cc63e36288a0275a37)